chore: move progress into task.go
