+++ api/task.go
@@ -169,6 +169,21 @@ type Task struct {
 	Progress Progress `jsonapi:"attr,progress"`
 }
 
+// Progress is a generalized struct which can track the progress of a task.
+type Progress struct {
+	// TotalUnit is the total unit count of the task
+	TotalUnit int64 `json:"totalUnit"`
+	// CompletedUnit is the finished task units
+	CompletedUnit int64 `json:"completedUnit"`
+	// CreatedTs is when the task starts
+	CreatedTs int64 `json:"createdTs"`
+	// UpdatedTs is when the progress gets updated most recently
+	UpdatedTs int64 `json:"updatedTs"`
+	// Payload is reserved for the future
+	// Might be something like {comment:"postponing due to network lag"}
+	Payload string `json:"payload"`
+}
+
 // TaskCreate is the API message for creating a task.
 type TaskCreate struct {
 	// Standard fields
